requirements-vercel.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# Database
SQLAlchemy==2.0.21

# Data Analysis (lightweight versions)
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2

# HTTP requests
requests==2.31.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7

# Flask-Migrate for database migrations
Flask-Migrate==4.0.5

# SocketIO for real-time features
Flask-SocketIO==5.3.6

# Environment variables
python-dotenv==1.0.0